Created Game class and moved 2 card methods there
